Move CoverInfoRelative/CoverInfo/CoverArt to track/coverinfo.h
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Created track/coverinfo.h containing CoverImageUtils, CoverInfoRelative,
CoverInfo, CoverArt, and their operators/helpers. These are track-domain
types that describe cover art metadata stored alongside a track; they do
not belong in library/.

Replaced library/coverart.h with a single-line compatibility redirect to
track/coverinfo.h. All existing #include "library/coverart.h" sites
continue to work without changes.

Updated track/trackrecord.h to include track/coverinfo.h directly instead
of library/coverart.h. This removes the last track/ → library/ include
dependency, removing the dependency that kept track/ headers coupled to library/.

The CoverInfo::loadImage() implementation remains in library/coverart.cpp,
which includes library/coverartutils.h. This is intentional: loadImage()
needs CoverArtUtils::extractEmbeddedCover(), which is a library-level
operation. The declaration is in track/, the heavy implementation stays
in library/ — the library module is allowed to depend on track.
