+++ src/track/trackrecord.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "library/coverart.h"
 #include "proto/keys.pb.h"
+#include "track/coverinfo.h"
 #include "track/cue.h"
 #include "track/keys.h"
 #include "track/keyutils.h"
